Standardize System2 on int per #2514.

+++ drake/systems/framework/basic_vector.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <limits>
 #include <stdexcept>
 #include <vector>
@@ -37,7 +38,7 @@ class BasicVector : public VectorInterface<T> {
     values_ = value;
   }
 
-  ptrdiff_t size() const override { return values_.rows(); }
+  int size() const override { return values_.rows(); }
 
   Eigen::VectorBlock<const VectorX<T>> get_value() const override {
     return values_.head(values_.rows());
